+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * refactor JAX-RS context
 * AbstractAPI is now deprecated
 * add structured logging
+* add orchestration module (#38)
 
 # Version 1.10
 * Update dependencies