+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * AbstractAPI is now deprecated
 * add structured logging
 * add orchestration module (#38)
+* add push service (#12)
 
 # Version 1.10
 * Update dependencies