+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Version 1.12
+* JAXRSContextImpl has to be normal component not Prod only (#42)
+
 # Version 1.11
 * refactor JAX-RS context
 * AbstractAPI is now deprecated